pyproject.toml
[tool.poetry]
name = "AFragmenter"
version = "0.0.1"
description = ""
authors = ["Stefaan Verwimp <stefaan.verwimp@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/sverwimp/AFragmenter"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

[tool.poetry.dependencies]
python = "^3.9"
python-igraph = "^0.11.6"
numpy = "^2.0.0"
matplotlib = "^3.9.2"
rich = "^13.8.1"
rich-click = "^1.8.3"
biopython = "^1.84"

[tool.poetry.scripts]
afragmenter = 'afragmenter.cli:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
